feat: Refactor RecentActivity component to fetch activities from database and move utility functions to a separate module
